Create services folder and reorganized files.
