+++ src/helpers/formatBigIntWithDecimals.ts
