+++ src/query/useSignGetSEP10cChallenge.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { SEP10cService } from "@/helpers/SEP10cService";
+import { SEP10cService } from "@/services/SEP10cService";
 import { ContractSigner } from "@/types/types";
 import { xdr } from "@stellar/stellar-sdk";
