+++ src/query/useTransfer.ts
@@ -2,7 +2,7 @@ import { XdrLargeInt } from "@stellar/stellar-sdk";
 import { useMutation } from "@tanstack/react-query";
 
 import { ScConvert } from "@/helpers/ScConvert";
-import { SorobanService } from "@/helpers/SorobanService";
+import { SorobanService } from "@/services/SorobanService";
 import { ContractSigner } from "@/types/types";
 
 type TokenTransferProps = {
